+++ nengo_dl/tests/test_simulator.py
@@ -1129,12 +1129,19 @@ def test_get_nengo_params(Simulator, seed):
 
 @pytest.mark.parametrize("progress", (True, False))
 def test_progress_bar(Simulator, progress):
+    net, _, p = dummies.linear_net()
+
     # note: ideally we would capture the stdout and check that output is
     # actually being controlled. but the pytest capturing doesn't work,
     # because it's being printed in a different thread (I think). so we just
     # check that the parameter works without error
-    with Simulator(nengo.Network(), progress_bar=progress) as sim:
+    with Simulator(net, progress_bar=progress) as sim:
         sim.run_steps(10, progress_bar=progress)
+        sim.loss(10, {p: lambda x: x}, progress_bar=progress)
+
+        if not sim.tensor_graph.inference_only:
+            sim.train(10, tf.train.GradientDescentOptimizer(0),
+                      objective={p: lambda x: x}, progress_bar=progress)
 
 
 @pytest.mark.training